Fix ogr-decode --where query syntax
Contrary to intermittent (which is set to either yes or no), the seasonal OSM tag is unset when the lake or stream is not seasonal.
An equality check against '' (empty string) in a where query in that case will evaluate to false.
Using the is operator instead, and checking directly against NULL, fixes that, the condition evaluates to true, and lakes and streams are correctly generated.

See https://gis.stackexchange.com/q/473101/190214 and https://stackoverflow.com/q/9581745/14909980